requirements.txt
Flask==2.2.2
gunicorn==20.1.0
google-cloud-storage==2.7.0
sendgrid==6.9.7